ai_ops_assistant/requirements.txt
langchain>=0.2.0
langgraph>=0.0.40
langchain-openai>=0.1.8
langchain-community>=0.2.0
streamlit>=1.33.0
requests>=2.31.0
python-dotenv>=1.0.1
duckduckgo-search>=5.3.0
